% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smooth_fdr}
\alias{smooth_fdr}
\title{Smooth FDR workhorse}
\usage{
smooth_fdr(probe, z, pos, chr, nulltype, nlambda, tol, maxit)
}
\description{
This is the workhorse of the smoothFDR package. The eponymous function
is just a wrapper.
}
