\name{smoothFDR-package}
\alias{smoothFDR-package}
\alias{smoothFDR}
\docType{package}
\title{
\packageTitle{smoothFDR}
}
\description{
\packageDescription{smoothFDR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{smoothFDR}
\packageIndices{smoothFDR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{smoothFDR}

Maintainer: \packageMaintainer{smoothFDR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
